+++ src/steps/stepInterfaces.ts
@@ -120,3 +120,7 @@ export interface ReturningStep extends BaseStep {}
 export interface ValuesStep extends BaseStep {
   returning(...items: (ItemInfo|ReturningItem|PrimitiveType)[]): ReturningStep
 }
+
+export interface DefaultValuesStep extends BaseStep {
+  returning(...items: (ItemInfo|ReturningItem|PrimitiveType)[]): ReturningStep
+}