+++ test/exported/insertPath/insert.test.ts
@@ -145,4 +145,20 @@ describe('INSERT Path', () => {
       expect(actual).toEqual(`INSERT INTO "table1"("col1", "col2", "col3") VALUES('A', DEFAULT, 'B');`)
     })
   })
+  describe('Insert with DEFAULT VALUES keyword', () => {
+    it(`Produces [INSERT INTO "table1"("col1") DEFAULT VALUES;]`, () => {
+      const actual = sql
+        .insertInto(table1, col1)
+        .defaultValues()
+        .getSQL()
+      expect(actual).toEqual(`INSERT INTO "table1"("col1") DEFAULT VALUES;`)
+    })
+    it(`Produces [INSERT INTO "table1"("col1", "col2", "col3") DEFAULT VALUES;]`, () => {
+      const actual = sql
+        .insertInto(table1, col1, col2, col3)
+        .defaultValues()
+        .getSQL()
+      expect(actual).toEqual(`INSERT INTO "table1"("col1", "col2", "col3") DEFAULT VALUES;`)
+    })
+  })
 })